requirements.txt
amqp~=5.1.1
asgiref~=3.6.0
billiard~=3.6.4.0
celery~=5.2.7
click~=8.1.6
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
Django~=3.2.7
django-celery-beat==2.6.0
django-timezone-field~=6.1.0
djangorestframework~=3.12.3
djangorestframework-simplejwt==5.3.1
drf-extensions==0.7.1
kombu~=5.3.1
postgres==4.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
psycopg2-pool==1.2
PyJWT==2.8.0
python-crontab~=3.0.0
python-dateutil~=2.8.2
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse~=0.2.4
tzdata==2024.1
vine~=5.0.0
wcwidth~=0.2.6

cryptography~=39.0.1
typing_extensions~=4.13.0
pip~=23.2.1
attrs~=23.1.0
distro~=1.9.0
pillow~=10.3.0
serialization~=1.2.3
redis~=5.0.4
pymongo~=3.12.3
SQLAlchemy~=2.0.35
isodate~=0.6.1
Brotli~=1.1.0
greenlet~=3.0.2
requests~=2.26.0
psycopg2~=2.9.10
PyYAML~=6.0.1
numpy~=1.26.2
cffi~=1.15.1
setuptools~=65.5.0
packaging~=23.2
zipp~=3.21.0
future~=0.18.3
uritemplate~=4.1.1
inflection~=0.5.1
pytz~=2023.3
django-filter~=2.4.0